requirements.txt
streamlit
fastapi
uvicorn
requests
python-dotenv
openai
langchain
